BREAKING CHANGE:
Reordered the fields of Graph per issue #53 in order to allow streaming implementations to read version info first in order to either fail fast or quirk based on the IR version and producer version used to produce the graph.

Note that this is a breaking change to (a) any serialized GraphProto's and (b) any code that used the GraphProto.producer_version, which is now encoded as a string given that we can't assume that all frameworks/runtimes that will emit ONNX adhere to the SemVer versioning scheme.