Complain about stale onnx_pb2 file.
I noticed some library code was importing onnx_pb2 incorrectly.  The
name of this file has changed, but the old source file was left around
because it was generated by protoc.  This leads to weird errors like
TypeError: 'GraphProto' object is not iterable because we're doing
a GraphProto test with two things named GraphProto, but not the same
thing.

Easiest to tell the user to remove the stale file manually.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>