Fix checker to know about versions.
Now all checker functions take a CheckerContext, which specifies what
the set of in-scope opset imports are.  We use this to decide which
version of a schema to pull out of the schema checker when doing
validation.

Public facing APIs default to the most recent operator set version
we know about.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>