Added test for 2D and 3D for average and max pool.
I implemented tests for 2D and 3D max and mean pooling,
comparing the results to a calculation in numpy.
1D pooling still had issues using the caffe2
backend so I left them out for the time being.
In addition, the testing with the caffe2
backend still fails as it uses ceil to calculate the
size of tensors instead of floor. The values
for the existing entries still match though.