Wrapper for vendor-specific ONNXIFI libraries
Add a library (onnxifi.dll/libonnxifi.so/libonnxifi.dylib), which finds and load ONNXIFI implementations in the system, and exposes all their backends under a single interface.
With this library, applications can just use

  onnxifi_load(ONNXIFI_LOADER_FLAG_VERSION_1_0, NULL, &library);

and get all ONNXIFI backends in the system.

The libraries are searched in the following directories:
- GNU/Linux: /usr/lib
- macOS: /opt/onnx/lib
- Windows: system directory (usually C:\Windows\System32)
- Default location can be overriden with CMake configuration-time variable ONNXIFI_SEARCH_DIR

Vendor-specific libraries must follow the rules below to be successfully found:
- Library filename must be onnxifi-${suffix}.dll (Windows), libonnxifi-${suffix}.dylib (Mac), or libonnxifi-${suffix}.so (Linux and other OSes)