Fix output_shape of a testcase for ConvTranspose (#1437)
* Fix output_shape of a testcase for ConvTranspose

As per the documentation of the operator, `output_shape` seems
to expect the same rank as other attributes such as `strides`
or `pads`. In other words, `output_shape` should not have
dimensions for batch size and number of channels. So this
changes `output_shape` from [1, 2, 10, 8] to [10, 8] for the
2D ConvTranspose.

Note that test_convtranspose_kernel_shape, the other test which
uses `output_shape`, is already OK.

* Update the document by gen_doc.py

* Update test coverage by stat_coverage.py

Note there are a few unrelated updates.

* Revert unrelated update in the previous commit