Tighten wording for operator sets regarding breaking changes
The wording regarding op set versioning is currently unclear regarding optional attributes. Although not stated in the document, it is typical in practice to update the version whenever attributes are added, and it is desirable for the benefit of model loaders to quickly ascertain their compatibility with the operators in the model. It would be bad for a model loader to think that it can successfully process a model only to yield incorrect results because it was unaware of a newly added attribute. Additionally, this will tighten wording to include tensor data type additions/removals.