Enhance shape inference for tile.
Currently tile's shape inference only gets the elem_type. This commit
enhances it to get the rank same as the input.
If the repeats input is specified in the model. We can also get the
exact size of each dim.

Also includes an extremely minor change to replace {} to {{}}
in shape_inference_test.cc to make it compile with c++11.

Changes to be committed:

    modified:   onnx/defs/tensor/defs.cc
    modified:   onnx/test/cpp/shape_inference_test.cc
    modified:   onnx/test/shape_inference_test.py