Remove GraphCall (#2964)
* Remove GraphCall
Most of the time, we only have one GraphCall inside the training
graph for calling `forward` of the model once. Thus, it seems more
natural to automatically add that `forward` into training graph.
This PR changes the IR for training and change training graph to
a partial graph. The full training graph is formed by merging
(partial) training graph and the inference graph. In addition,
GraphCall is removed because it will be rarely used.

* Address comments