Fixes a missing attribute in function implementation of CastLike (#5246)
### Description
The function implementation of CastLike does not propagate the attribute
*saturate*. This PR fixes it and make the changes in the reference
implementation to check the function implementation is working. This PR
fixes two issues:

1. function implementation of CastLike (19)
2. reference implementation working context dependent function

By default, the reference implementation choose a kernel if there is
one, even if the operator defines an implementation. The PR introduces a
mechanism to force the use of the kernel and expand the operator by its
function. It delays that when the function is context dependant. It is
then done when the node is executed and its inputs known.

---------

Signed-off-by: xadupre <xadupre@microsoft.com>