Allow to override cmake options from cmd options, external projects (#5355)
### Description
- Allow to override CMAKE_CXX_STANDARD and
CMAKE_POSITION_INDEPENDENT_CODE

### Motivation and Context
- CMAKE_POSITION_INDEPENDENT_CODE is not required for cases when static
library is compiled into executable. So, we need to a way to disable
`CMAKE_POSITION_INDEPENDENT_CODE=ON`
- Protobuf depends on ABSL library, which requires CXX17. So, when we
compile against latest versions of protobuf and ABSL, we need to set
CXX17, but ONNX overrides these values to CXX14. So, we need an option
to override the project default values.

Signed-off-by: Ilya Lavrenov <ilya.lavrenov@intel.com>
Co-authored-by: Chun-Wei Chen <jacky82226@gmail.com>
Signed-off-by: Xavier Dupre <xadupre@microsoft.com>