Fixes math norm rendering in docs (#5391)
Fixes the math norm rendering issue in the official docs. 

### Description
Fixes a bug with latex math norm rendering which occurs in the official
documentation and the github markdown rendering. Changes usage of
`\norm{}` to `\lVert \rVert`.

### Motivation and Context

See images below for examples of the errors and fix in the official
documentation at
https://onnx.ai/onnx/intro/concepts.html#input-output-node-initializer-attributes
and in the github markdown preview.

#### Error on Github 


![norm_error](https://github.com/onnx/onnx/assets/50355587/45ed3839-b3ad-4a74-9130-96b3630844e5)

#### Fix on Github 


![norm_fix](https://github.com/onnx/onnx/assets/50355587/0ba86c7f-a735-444d-a564-9f5062d6c7b8)

#### Error on Official Docs 

See :
https://onnx.ai/onnx/intro/concepts.html#input-output-node-initializer-attributes

![norm_error_official](https://github.com/onnx/onnx/assets/50355587/f3f41a59-49f2-4d62-bbb1-7180d1e0b36d)

#### Note 

This fix has only been tested with the markdown preview due to errors
building onnx from source to preview the sphinx docs.

Signed-off-by: Gabriel Dennis <gabrieldennis17@gmail.com>
Co-authored-by: Xavier Dupré <xadupre@users.noreply.github.com>