Solve merge queue timeout issue: accelerate windows CI (#5431)
### Description
<!-- - Describe your changes. -->
- I noticed typically py38_onnx_ml pipeline was very slowly created...
Using newer version of Python 3.9 might help. If it is still too slow,
we need to review the build time for build internal Protobuf and see
whether we can make it faster.
- Set maxParallel as 6 to ensure every pipelines on Windows run at the
same time.
- Make heavy release CIs cancel-in-progress.


### Motivation and Context
<!-- - Why is this change required? What problem does it solve? -->
<!-- - If it fixes an open issue, please link to the issue here. -->
Merge queue automatically removed PR, because some of pipelines (mainly
py38_onnx_ml on Windows) are too slow and hit the timeout for merge
queue. We should solve it in both ways:
- Increase the timeout for merge queue (just increased from 60 to 90
mins)
- Accelerate the slow Windows CI if possible

The PR is doing the later one. The former one needs to be done by the
maintainer. cc @prasanthpul.

---------

Signed-off-by: Chun-Wei Chen <jacky82226@gmail.com>