Allow choosing serializers based on file extension (#5347)
### Description

1. Allow inferring serializers based on file extension
2. Update the `ProtoSerializer` protocol to support one format per class
and include a new field to specified supported extensions. **This
changes the interface of `ProtoSerializer` but that's ok since the
interface was introduced in 1.15dev**

### Motivation and Context

Now users can do

```python
onnx.save_model("model.textproto")
```

without having to specify format.

When a format is specified, it takes precedence over the file extension.

---------

Signed-off-by: Justin Chu <justinchu@microsoft.com>