reference implementation of pool ops to support dilation (#5250)
### Description
With reference implementation, dilation is not supported for AvergePool
and LPPool. This PR provide such support

### Motivation and Context
To improve test coverage
https://github.com/onnx/onnx/issues/4849
https://github.com/onnx/onnx/issues/5161

The next step will be to replace maxpool with the new reference
implementation. I still need time to figure out how to support argmax
for maxpool with the new reference implementation therefore need another
PR for this work.

---------

Signed-off-by: Liqun Fu <liqfu@microsoft.com>