Support selective inlining in function inliner (#5515)
### Description
This extends the function-inliner to support selective function
inlining. The caller can indicate which functions are to be inlined.

Consider an example where the main-graph calls functions f1 and f2.
Suppose further that f2 calls f1. If we ask for f1 to be inlined, but
not f2, then the call to f1 in both main and f2 will be inlined. But f2
will not be inlined into main.

This introduces one significant change to the inlining logic: now, the
call-site to be inlined may be one inside a FunctionProto (instead of a
GraphProto). This has some implications, namely,
full-context-information (types of input parameters or values of
attribute parameters)_may not be known at inlining time. The inlining is
still doable. However, more advanced features like version-conversion,
when inlining, cannot be supported.

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>