Create the onnxtxt serialization format (#5524)
Add support for serializing to and deserializing from the ONNX textual
representation, and register it under the `onnxtxt` file extension and
format.

Users can now save a model to the textual representation with
`onnx.save_model(model, "model.onnxtxt")`

Note that the textual format has known limitations, including (1) it
doesn’t serialize tensor protos or initializers and (2) bugs like #5102.
We will call this out in a documentation in a separate PR.

---------

Signed-off-by: Justin Chu <justinchu@microsoft.com>
Signed-off-by: Justin Chu <justinchuby@users.noreply.github.com>