Rewrite AffineGrid function body to make it pass shapeInference (#5501)
### Description
Function body of AffineGrid is written to pass shape inference. It
treats the 2d case as a special from of 3d. ops are carefully used to
ensure there is not if-condition specific elements.

### Motivation and Context
ONNX shape inference is run for both then and else subgraphs regardless
of condition. This may fail if shape inference is "if condition"
specific. This issue is fundamental to ONNX shape inference. We have
several strategies to handle it
https://github.com/onnx/onnx/issues/4986#issuecomment-1679265312. One
solution is to construct the function body to remove "if-condition"
specific part.

---------

Signed-off-by: Liqun Fu <liqfu@microsoft.com>