Enable pylint checks with Ruff and remove pylint from lintrunner (#5589)
### Description

Remove the `pylint` linter in favor of the `PL` rules provided by ruff.
This change removes all of the `pylint: disable` directives in comments
and added `noqa` directives accordingly for ruff.

- Remove auto format at VS Code save: Removed because they significantly
slow down file save when doing a bulk find and replace. Since we have
introduced lintrunner autofix, developers can use a single command for
formatting which should be easy enough.

### Motivation and Context

Ruff has implemented most of the pylint checks we care and it runs much
faster.

---------

Signed-off-by: Justin Chu <justinchu@microsoft.com>