Getting onnx to treat `inf/-inf` as float literals. (#5528)
### Description
<!-- - Describe your changes. -->
onnx seems to ignore certain possibilities when it comes to otherwise
valid IEEE float values. So far this change just adds functionality to
the parser to recognize `inf` or `-inf` as floats. However, this seems
to point towards more issues. Should onnx recognize any string
parse-able by `std::stof` as a float literal?

### Motivation and Context
<!-- - Why is this change required? What problem does it solve? -->
<!-- - If it fixes an open issue, please link to the issue here. -->
The current state of this PR does address #5102, although not in a very
elegant fashion.

---------

Signed-off-by: Clif Houck <me@clifhouck.com>