Enable empty list of values as attribute (#5559)
This PR makes two fixes:

* The existing checker does not allow an attribute to have an empty list
(of ints, or floats, or strings, etc.) as a value. The PR removes this
check. It is better for op-specific checks to check for this where
required.
* There are similar checks in the shape-inference method of the Constant
op, which are also removed.

This allows a Constant op to construct an empty 1D tensor of ints or
floats, for example, conveniently.

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>