Drop "one of" default attribute check in LabelEncoder (#5673)
### Description
Drop "one of" checking on the default value in the `LabelEncoder`.

### Motivation and Context
When implementing `LabelEncoder` in onnxruntime (for the upcoming
ai.onnx.ml opset 4), it became clear that the implemented type inference
for LabelEncoder would not work well in practice - specifically, the
type inference pass ensures that only one of the default_* attributes is
set. Since some of these are optional with default values, this is not
very workable in practice. This PR addresses this by dropping this
check. Notice that this was [not
checked](https://github.com/onnx/onnx/blob/f556cffc40dc769a2c8e3bb5fd9a90f1e7fbb7eb/onnx/defs/traditionalml/old.cc#L277-L338)
in previous opsets anyway.

---------

Signed-off-by: Aditya Goel <agoel4512@gmail.com>