TreeEnsemble base values for the reference implementation (#5665)
This PR fixes one line failing on #5569. What follows comes from this PR
description.

### Description

In the reference implementation of TreeEnsembleRegressor when a value is
provided for the base_values argument, this value replaces any
prediction. This can't be what's intended and does not match
onnxruntime. Instead, the base_value should be added to the prediction
after applying any aggregation.

### Motivation and Context

We are exporting regression trees into ONNX that have a non-zero
base_value as the baseline prediction for the tree. Prediction works as
expected in onnxruntime but not in the reference implementation. I
believe this is an oversight and propose the fix (plus tests) below. I
also think the documentation should be more explicit about what the
base_values argument does.

---------

Signed-off-by: Corwin Joy <corwinjoy@gmail.com>
Signed-off-by: Xavier Dupre <xadupre@microsoft.com>
Co-authored-by: Corwin Joy <corwin.joy@gmail.com>
Co-authored-by: G. Ramalingam <grama@microsoft.com>