Parser/printer support external data format (#5688)
### Description
Improve the implementation of ONNX textual printer / parser to handle
external data format

### Motivation and Context
Recently the modern models get larger and larger. Models with external
data become common.
The ONNX textual format is good for human writing and reading. But is
dose not implement the
external data format, yet. This PR implement it.

---------

Signed-off-by: yoco <yoco@skymizer.com>
Signed-off-by: Justin Chu <justinchuby@users.noreply.github.com>
Co-authored-by: yoco <yoco@skymizer.com>
Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>