Fix endianness conversion in numpy_helper.to_array() (#5904)
### Description
This removes the invocation of `convert_endian()` from
`numpy_helper.to_array()` in favor of another implementation, since
convert_endian() modifies the TensorProto.

### Motivation and Context
This change was needed due to observed issues when trying to load an
ONNX model that was exported on a s390x system via tf2onnx. Debugging
showed that the TeensorProto was being modified by endianness conversion
operations, which led to model incorrectness under certain
circumstances.

Fixes #5901.

Signed-off-by: Tabari Alexander <tehbone@gmail.com>