Add valueinfos field to FunctionProto (#5903)
### Description

* Extend FunctionProto by adding a field to store ValueInfos for
variables in the function.
* Extend the helper function and parser to support this.

### Motivation and Context

This enables type/shape information to be stored for variables in a
FunctionProto, just like in a GraphProto. Note that this is optional,
just like in a graph. This field may store type/shape info for the
inputs and outputs of a function as well. (Note that, unlike in a graph,
the input/output field in a FunctionProto do not have explicit
types/shapes, they are just strings, not ValueInfo.)

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>