Add proto support for overloaded functions (#5011)
A minimal and simple extension to the proto definitions to allow
multiple (overloaded) functions in its model-local function list with
same (domain, name).

Model-local functions in ONNX provide a default-implementation for
certain ops used in the model (expressed in terms of other primitive
ops). This extension will allow us to provide different
default-implementations for different calls to the same op (which might
differ in the types of input-arguments or in the values of
attribute-parameters).

* Update proto
* Update helper.py constructors
* Update parser
* Update shape inference implementation
* Update in-memory IR
* Add test-case
* Minimal extensions in inliner to support this

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>
Signed-off-by: Abhishek-TyRnT <kulkarnia168@gmail.com>