Replace rank inference with shape inference for Einsum op (#6010)
In the development of ONNX Runtime, we need know the output shape of
each Op node for statical graph compilation. However, we found that we
could use onnx shape inference to achieve almost all output shapes
except the output shape of Einsum. In `onnx/defs/math/defs.cc`, we found
that there was only Rank Inference function for Einsum instead of Shape
Inference.
Given the results of equation parsing and the input shapes of the Einsum
operation, we can easily infer the output shape. As such, we have
expanded the rank inference to include shape inference for Einsum.

---------

Signed-off-by: peishenyan <peishen.yan@intel.com>
Co-authored-by: G. Ramalingam <grama@microsoft.com>