Check graph outputs are defined (#6083)
The current checker does not check that all graph outputs are defined.
This PR adds such a check. (See issue
https://github.com/onnx/onnx/issues/6040 ).

Note: We probably also want a separate check to ensure that a
graph-input is not used as a graph-output. But leaving that for a
separate PR, since it is a distinctly different check (and perhaps might
require a different discussion).

Redo of PR #6070

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>
Signed-off-by: Linsho Kaku <linsho@preferred.jp>