Corrected the ReduceSumSqure op to match the reference (#6121)
### Description
This PR aligns the op implementation for `ReduceSumSquare18` when `axes`
are not specified and `noop_with_empty_axes != 0` with that of the
reference.

### Motivation and Context
Current implementation of `ReduceSumSquare18` when axes are empty and
`noop_with_empty_axes != 1` squares the input data and then returns it,
when according to the reference it should just return the input data
with no changes.

This addresses issue #6103

Signed-off-by: Aman K Shihab <amanshihab276@gmail.com>
Signed-off-by: Linsho Kaku <linsho@preferred.jp>