Add alternative approach in ONNX converter section (#6188)
### Description
This adds a short section explaining an alternative approach to
dedicated converter libraries for library authors looking to implement
ONNX export capability for their library. This approach leverages a
standardised protocol (in this case the [Array API
standard](https://data-apis.org/array-api/latest/)) and an ONNX based
implementation of it (in this case
[ndonnx](https://github.com/Quantco/ndonnx)).

### Motivation and Context
This adds some broader context to the different ways libraries may want
to explore ONNX export beyond the traditional approach with dedicated
converter libraries.

Signed-off-by: Aditya Goel <agoel4512@gmail.com>