Upgrade reuse to v4.0.0 (#6216)
### Description
Upgrade Reuse to v4.0.X

### Motivation and Context
According to the release notes: 

"This release of REUSE implements the new [REUSE Specification
v3.2](https://reuse.software/spec-3.2). It adds the REUSE.toml file
format as a replacement for .reuse/dep5. The new format is easier to
write and parse, is better at disambiguating certain corner cases, and
is more flexible for customisation and future additions."

---------

Signed-off-by: Andreas Fehlner <fehlner@arcor.de>