Use MACOSX_DEPLOYMENT_TARGET=12.0 for macOS wheels (#6242)
### Description
It seems the latests wheels aren't compatible with macOS 11 for some
reason.
We can't test on macOS 11 anymore & macOS 11 is EOL.
Use MACOSX_DEPLOYMENT_TARGET=12.0 and move the tests to the lowest
versioned runner available.

Note: the delocate tool did not find anything wrong with the target
being 11.0 on the 1.16.1 wheel. It should have complained if using
symbols only compatible with 12.0+ so it seems
MACOSX_DEPLOYMENT_TARGET=11.0 was indeed respected and the root cause of
the issue is still unknown.

### Motivation and Context
#6191

Signed-off-by: mayeut <mayeut@users.noreply.github.com>