ExternalData - On Windows document proper offset for memory mapping support (#6247)
### Description
On Windows, it is incorrect to say that offset needs to be multiple of
page size. Instead offset needs to be multiple of VirtualAlloc
allocation granularity.

You can see an error returned in ORT code if this is not the case
https://github.com/microsoft/onnxruntime/blob/4e75605eec985e579ee1e8db9a2bb2fd441a4837/onnxruntime/core/platform/windows/env.cc#L453

### Motivation and Context
We debugged an issue with ORT team where memory mapping was not working.
Even when aligned to page_size, on Windows, memory mapping did not work
until aligned to 64kb.
See
[MapViewOfFile](https://learn.microsoft.com/en-us/windows/win32/api/memoryapi/nf-memoryapi-mapviewoffile)

Related: #6226

Signed-off-by: ivberg <ivberg@microsoft.com>