Update broken URLs (#6255)
### Description
Fix some broken URLs that failed test when I tried to create rel-1.16.2
branch.

### Motivation and Context
* https://docs.scipy.org/doc/numpy/reference/arrays.indexing.html
* Automatically redirects to
https://numpy.org/doc/stable/reference/arrays.indexing.html which now
404's
* Based on web.archive.org search,
https://numpy.org/doc/stable/reference/routines.indexing.html has the
same information as the old URL so updating to use that link instead.
*
https://github.com/pytorch/pytorch/blob/master/caffe2/python/net_drawer.py
* Automatically redirects to main, but that file no longer exists in
that branch. The last tag with the file was
https://github.com/pytorch/pytorch/blob/v2.3.1/caffe2/python/net_drawer.py.
* Based on the v2.3.1 version, I tried searching for a couple of the
method names but got no hits. So I think the file was deleted rather
than renamed. Updated the URL in the comment to the last version
available.

---------

Signed-off-by: Charles Volzka <cjvolzka@us.ibm.com>