Combine different release pipelines by the use of reusable workflows (#6277)
### Description
<!-- - Describe your changes. -->

This pull request should be the start of a pipeline (current status of
the considerations under https://github.com/onnx/onnx/pull/6246) that
automatically generates a whls after tagging, signs it and publishes it
directly to Pypi to improve supply chain security. As this is all too
much at once, the changes will be introduced and tested gradually.

Changes related to this PR:
Reusable Workflows are used to merge the individual OS release builds
together (we need one process as we want to have all artifacts in one
pipeline, so there is not need to download them manuelly). The workflows
are currently not really "reusable", as we more or less use pypi
creditionals in the workflow, among other things. I reused our testpypi
weekly repo as I didn't want to change our behavoir for pypi and
testpypi.

Maybe it would make sense to adapt the process right away so that you
can trigger our "Weekly" manually via the Github interface for testing
and don't have to wait a week ;-)

#### Key points for the review:
* How do we want to configure concurrency
* What concurrency-groups do we need?
* Where do we need "cancel-in-progress"?

#### Not part of the pull request:
* Extract publish whl to pypi/testpyp
* Introduce trusted publishing
* ...

### Motivation and Context
* see https://github.com/onnx/onnx/pull/6246

---------

Signed-off-by: Andreas Fehlner <fehlner@arcor.de>
Co-authored-by: Justin Chu <justinchuby@users.noreply.github.com>
Signed-off-by: Linsho Kaku <linsho@preferred.jp>