Fix missing secrets for publishing of onnxweekly  (#6321)
### Description
After the release-pipeline upgrade
(https://github.com/onnx/onnx/pull/6277)
the publishing the weekly does not work any more as the secrets are not
available at the resuable workflows (
https://github.com/onnx/onnx/actions/runs/10551283959/job/29228497156 )

Link for the option I used for the secrets:

https://github.blog/changelog/2022-05-03-github-actions-simplify-using-secrets-with-reusable-workflows

### Motivation and Context

In my opinion, we no longer need these secrets at the end. 

In the future, the publishing of the Onnx weekly builds will be solved
via “trusted publishing” outside the individual OS release pipelines.
This is already included in create_release.yml. Currently, the previous
weeklys are published at pypi.org, in the test via trusted publishing at
test.pypi.org

After a successful schedule test (next Monday), we could adjust the
publishing. a) e.g. throw out the upload steps with twine and adjust the
target of separate steps... b) or leave it as it is for now...

Signed-off-by: Andreas Fehlner <fehlner@arcor.de>