Extend printer and parser to support invalid identifiers (#6346)
Many onnx models used in practice use tensor names that are not valid C
identifiers (as required by the ONNX standard). This PR extends the
printer and parser to support such identifiers.

(It may be useful to relax the ONNX standard's requirements in this
regard, but that's a separate issue, requiring some community
discussion.)

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>
Signed-off-by: Linsho Kaku <linsho@preferred.jp>