Support node labels in parser and printer (#6349)
### Description

1) Adds support for node label in parser and printer
2) Adds some status-checks that were missed in recent addition to
support invalid identifiers.

Node labels are now supported with the following syntax:

```
   [a_node_label] y = MatMul (x, w)
```

However, we have a couple of options to choose from here:

(a) The ONNX spec says that the node label must be a valid identifier.
So, we can allow for a valid identifier between the `[...]`, relying on
a quoted string inside if the user wants to use something that is not a
valid identifier.

(b) Alternatively, since `[` and `]` act as end-markers, like quotes, we
can allow any string in-between, that is taken to be the node label.
(Users will need to use an escape backslash if they want `]` in the
label.)

Option (b) is better if users want to use non-identifier labels, eg.,
numbers, etc. I am inclined to go with (b), even though it is extra
code. However, it won't be friendly to the use of white-space inside if
users want to use identifier labels. Eg., `[ mylabel ]` will include the
blank spaces in the label.

The implementation currently does only (a), which is simpler.

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>
Co-authored-by: Andreas Fehlner <fehlner@arcor.de>
Signed-off-by: Linsho Kaku <linsho@preferred.jp>