clarify clip with min>max (#6395)
### Description
Added the handling of min > max to the clip operator. It was previously
unspecified how to handle that case, now the specs are more precise and
match python/pytorch.

No implementation changes were necessary in the reference implementation
as it already used the python clip operator.

### Motivation and Context

Some users asked about this corner case, and the present PR aims to
specify precisely how this corner case is handled. It now matches
Pytorch & Python. Tensorflow is unspecified in this respect, so no
conflicts are introduced.

Background issue is here: https://github.com/onnx/onnx/issues/6165

---------

Signed-off-by: Alexandre Eichenberger <alexe@us.ibm.com>
Signed-off-by: Linsho Kaku <linsho@preferred.jp>