Support official Python launcher on Windows (#6407)
### Description

This change modifies the mypy protoc plugin batch file to attempt to use
the `py` executable if the `python` executable isn't found.

### Motivation and Context

The full context and motivation is covered in
https://github.com/onnx/onnx/issues/6406, but the summary is that
different installers on Windows give the Python executable different
names (`python` or `py`) and without this change this script blocks ONNX
installation on Arm-based PCs, or any other Windows machines that use
the official Python dot org installer.

Signed-off-by: Pete Warden <pete@usefulsensors.com>
Co-authored-by: Pete Warden <pete@usefulsensors.com>