Fix parser to handle empty optional parameters in edge cases (#6427)
Fix parser to handle empty optional parameters in edge cases. Currently,
it does not handle a missing optional parameter in the first position.

There are a few edge cases. We have to handle both the original way of
specifying a missing optional parameter (with nothing followed by a
comma, see examples below) and the added new way we have now with
support for quoted identifiers (using an empty string "").

Examples:
* `Op()` has no operands
* `Op(,x)` has two operands, the first being empty.
* `Op("")` has one operand, which is an empty string.
* `Op(,)` has one operand, which is an empty string.

Thus, we allow a trailing comma after a non-empty identifier with no
effect.
* `Op(x,)` has one operand, which is 'x'.

---------

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>