Fix getsize bug in extract_model (#6496)
### Description
<!-- - Describe your changes. -->

Always use `infer_shapes_path` to infer the model.

### Motivation and Context
<!-- - Why is this change required? What problem does it solve? -->
<!-- - If it fixes an open issue, please link to the issue here. -->

In the previous version, the os.path.getsize() function call can only
count the size of the onnx file, not the external data.
For example, an onnx file is only 1.2M but the external data is 3.3G.
`infer_shapes` will return an empty model.
```
1.2M model.onnx
3.3G model.onnx_data
```

---------

Signed-off-by: Po-Chu Hsu <tonypottera@gmail.com>