Fix `MAXIMUM_PROTOBUF` size (2GiB, not 2GB) (#6556)
### Description
While saving a model that was just barely under the 2GiB limit (using
Optimum), I noticed that it was saved with the external data format,
even though it should be able to be saved in a single file.

After further investigation, it was related to [these
lines](https://github.com/huggingface/optimum/blob/65a8a94adaf136dd677d28cfc837c0acfe993031/optimum/onnx/graph_transformations.py#L135-L138),
where the value of `onnx.MAXIMUM_PROTOBUF` is used to determine whether
to use the external data format. This value is hardcoded to 2GB
(=`2*10^9` bytes), even though it should be 2GiB (=`2*1024^3`),
according to the [protobuf
documentation](https://protobuf.dev/programming-guides/proto-limits/#:~:text=form%20must%20be-,%3C2GiB%2C,-as%20that%20is):

![image](https://github.com/user-attachments/assets/6b59ccaa-2099-4d4e-9a34-7c79ba8dce8e).

Unless there is a specific onnx specification which limits this to 2GB
(and not 2GiB), this PR may potentially fix a problem which has existed
for a long time.


### Motivation and Context
<!-- - Why is this change required? What problem does it solve? -->
<!-- - If it fixes an open issue, please link to the issue here. -->

Allows certain models to be saved as a single `.onnx` file without the
external data format, which were previously saved separately due to this
misconfiguration.

Signed-off-by: Joshua Lochner <admin@xenova.com>