Use rank in data propagation (#6557)
### Description

Redo of #6553 (due to DCO/signoff issues)

Extend the implementation of partial data propagation to make use of the
rank of a tensor, if known.

### Motivation and Context

This enables more precise shape inference, especially in the presence of
dynamic shapes involving some unknown dimensions. Generated models often
use computed shapes that mix statically known and statically unknown
dimensions (for example, in an expression such as `Concat (BatchSize,
SequenceLength, OtherKnownDimensions)`). Even if `BatchSize` and
`SequenceLength` are unknown, the other dimensions may be known, and it
useful to track them to enable certain optimizations.

Signed-off-by: Ganesan Ramalingam <grama@microsoft.com>