Fix reference implementation for TopK (#6593)
### Description
This PR fixes problems in the reference implementation of TopK in the
following cases by rewriting the tiebreaker mechanism and unifying logic
for all cases.

### Motivation and Context
The ONNX [documentation](https://onnx.ai/onnx/operators/onnx__TopK.html)
states:
> Given two equivalent values, this operator uses the indices along the
axis as a tiebreaker. That is, the element with the lower index will
appear first.

This was not the case currently as the `flip` after the `argsort`
changes the order of the same-valued elements.
Additionally the logic around 2d shaped inputs introduces a bug for
unsigned inputs as the negation overflowed.

This mismatch between the reference implementation and `onnxruntime` was
found by running the
[array-api](https://data-apis.org/array-api/latest/) test suite through
[ndonnx](https://github.com/Quantco/ndonnx) on both.

---------

Signed-off-by: neNasko1 <nasko119@gmail.com>
Co-authored-by: Andreas Fehlner <fehlner@arcor.de>