Allow scale type in QuantizeLinear and DequantizeLinear to be different from the high-precision type (#6531)
### Description
Allow the scale type in Q/DQ ops to be different from the high-precision
type.

- For Q, an additional optional attribute `precision` is introduced to
control the precision of the division operation. The default precision
is `y_scale` type.
- For DQ, an additional optional attribute `output_dtype` is introduced.
The default is to set the output type to `x_scale` type. The output type
is also the precision of the multiplication operation. The user can
always add a cast node after DQ to cast to another desired type.

### Motivation and Context
A motivating example is when the high-precision data type of Q/DQ is
fp16, sometimes it can be beneficial to have fp32 scale to preserve more
precision.

Another example is for [OCP’s MX-compliant
formats](https://www.opencompute.org/documents/ocp-microscaling-formats-mx-v1-0-spec-final-pdf)
the scale type is always E8M0.
In such cases, we would like to loosen the current constraint in the
ONNX spec that the scale type and the high precision type should always
be the same.

---------

Signed-off-by: Yuan Yao <yuanyao@nvidia.com>
Co-authored-by: Andreas Fehlner <fehlner@arcor.de>