Add cmake dependency only when system cmake is not available
Instead of unconditionally pulling `cmake` PyPI package in, check
whether `cmake` is available first and add the dependency only when
it is not available.  This is the same approach as used
by the `scikit-build-core` package, and it improves portability,
especially given that CMake frequently requires downstream patching
to work.

In this version of the patch, I've created a minimal PEP517 build
backend that inherits the functions provided by setuptools, and adds
the CMake dependency when necessary.  This avoids using deprecated
setuptools `setup_requires` argument, and it avoids adding CMake
dependency when building source distributions.

Signed-off-by: Michał Górny <mgorny@gentoo.org>